+++ nextgen/task-expander/tests/token_expander_test.rs
@@ -1,58 +1,10 @@
+mod utils;
+
 use moon_common::path::WorkspaceRelativePathBuf;
-use moon_common::Id;
 use moon_config::{InputPath, LanguageType, OutputPath, ProjectType};
-use moon_project::{FileGroup, Project};
-use moon_task::{Target, Task};
 use moon_task_expander::TokenExpander;
-use rustc_hash::FxHashMap;
 use starbase_sandbox::{create_empty_sandbox, create_sandbox, predicates::prelude::*};
-use std::path::Path;
-
-fn create_project(workspace_root: &Path) -> Project {
-    let source = WorkspaceRelativePathBuf::from("project/source");
-
-    Project {
-        id: Id::raw("project"),
-        root: workspace_root.join(source.as_str()),
-        file_groups: FxHashMap::from_iter([
-            (
-                "all".into(),
-                FileGroup::new_with_source(
-                    "all",
-                    [
-                        source.join("*.md"),
-                        source.join("**/*.json"),
-                        source.join("config.yml"),
-                        source.join("dir/subdir"),
-                    ],
-                )
-                .unwrap(),
-            ),
-            (
-                "dirs".into(),
-                FileGroup::new_with_source(
-                    "dirs",
-                    [
-                        source.join("other"),
-                        source.join("dir/*"),
-                        source.join("**/*.md"),
-                    ],
-                )
-                .unwrap(),
-            ),
-        ]),
-        source,
-        ..Project::default()
-    }
-}
-
-fn create_task() -> Task {
-    Task {
-        id: Id::raw("task"),
-        target: Target::new("project", "task").unwrap(),
-        ..Task::default()
-    }
-}
+use utils::{create_project, create_task};
 
 mod token_expander {
     use super::*;